pyproject.toml
[project]
name = "trm-agent"
version = "0.1.0"
description = "TRM (Tiny Recursive Model) aware language model training for tool-calling"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "sentencepiece>=0.2.1",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["trm_agent"]

[dependency-groups]
dev = []
